feat(core): workspace skeleton + extract atomic-write (KJC-TSK-0511)
PR1 of the @karajan/hu-board decoupling. Drops `packages/core` into the
workspace and lifts the first shared module (`atomic-write`) so both the
CLI and the board import it from `@karajan/core/atomic-write` instead of
cross-package relative paths.

- New private workspace `@karajan/core` with `./atomic-write` subpath
  export, a vitest config, and a smoke test covering both the async and
  sync writers.
- Root `package.json` `workspaces[]` adds `packages/core` alongside
  `packages/ai-trash`; `npm install` symlinks `node_modules/@karajan/core`
  on its own.
- `src/utils/atomic-write.js` becomes a re-export shim so the five CLI
  callers (`session/store`, `brain/standby-store`, `plan/plan-store`,
  `utils/run-registry`, `commands/plan/share`) stay untouched this PR.
- `packages/hu-board/src/plan-mutations.js` switches its import from
  `../../../src/utils/atomic-write.js` to `@karajan/core/atomic-write`.

Verified: `vitest run` on the 6 atomic-write consumer suites (41 passing),
the 3 hu-board mutation suites (60 passing), and the new core suite
(2 passing).
