+++ packages/core/src/index.js
@@ -1,0 +1,6 @@
+// Barrel export for @karajan/core. Subpath imports
+// (e.g. `@karajan/core/atomic-write`) are the preferred shape — this
+// barrel only exists so a single `import { writeJsonAtomic } from
+// "@karajan/core"` keeps working for callers that want the whole API.
+
+export * from "./atomic-write.js";
